내용 병합 관련
UserInform class 내용을 Account class에 병합
