+++ src/main/java/com/team04/musiccloud/auth/AccountRepository.java
@@ -12,6 +12,8 @@ public class AccountRepository {
     private Map<String, Account> accounts = new HashMap<>();
     private Random random = new Random();
 
+
+    //광: 암호화 한 후 데이터에 세이브한다.
     public Account save(Account account) {
         account.setId(random.nextInt());
         accounts.put(account.getEmail(), account);
